requirements.txt
streamlit
openai
pymupdf
python-docx
mail-parser